+++ execution/expiring_orders.go
@@ -1,4 +1,4 @@
-package matching
+package execution
 
 import (
 	types "code.vegaprotocol.io/vega/proto"
